requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
mediapipe>=0.10.0
Pillow>=10.0.0
scipy>=1.11.0
google-generativeai>=0.3.0
